+++ src/mainWindow.cpp
@@ -229,7 +229,8 @@ void MainWindow::populateInterfaceComboBox()
 	la::avdecc::networkInterface::enumerateInterfaces(
 		[this](la::avdecc::networkInterface::Interface const& networkInterface)
 		{
-			if (networkInterface.type != la::avdecc::networkInterface::Interface::Type::Loopback && networkInterface.isActive)
+			// Only display Ethernet interfaces
+			if (networkInterface.type == la::avdecc::networkInterface::Interface::Type::Ethernet && networkInterface.isActive)
 			{
 				_interfaceComboBox.addItem(QString::fromStdString(networkInterface.alias), QString::fromStdString(networkInterface.name));
 			}
@@ -533,6 +534,9 @@ void MainWindow::connectSignals()
 		{
 			LOG_HIVE_WARN("Failed to check for new version: " + reason);
 		});
+
+	auto* refreshController = new QShortcut{ QKeySequence{ "Ctrl+R" }, this };
+	connect(refreshController, &QShortcut::activated, this, &MainWindow::currentControllerChanged);
 }
 
 void MainWindow::showChangeLog(QString const title, QString const versionString)